Move now-passing test from under failing; update release notes
